debug

+++ src/inference/inference.cu
@@ -422,8 +422,7 @@ __global__ void kernel_standard_tiled_gemm(
 
         // Compute partial sums
         for (int i = 0; i < TILE_SIZE; ++i) {
-            value += __half2float(X_shmem[threadIdx.y * TILE_SIZE + i]) *
-                     __half2float(T_shmem[i * TILE_SIZE + threadIdx.x]);
+            value += X_shmem[threadIdx.y * TILE_SIZE + i] * T_shmem[i * TILE_SIZE + threadIdx.x];
         }
     }
